+++ packages/react/src/components/elements/repeaterGroup/wrapper.tsx
@@ -1,5 +1,5 @@
 import { Button } from "@/components/ui/button";
-import { ReactNode } from "react";
+import { ReactNode, useEffect } from "react";
 import {
   ArrayPath,
   FieldArray,
@@ -27,6 +27,14 @@ function RepeaterGroupWrapperComponent<T extends FieldValues>({
     control,
     keyName: "_key",
   });
+
+  useEffect(() => {
+    if (field.fields.length) {
+      return;
+    }
+
+    field.insert(0, {} as FieldArray<T, ArrayPath<T>>);
+  }, [field]);
 
   return (
     <div className="w-full flex flex-col gap-2 border border-lfui-border rounded-md border-dashed p-2">