+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lezzform/react
 
+## 0.2.3
+
+### Patch Changes
+
+- repeaterGroup init item
+
+## 0.2.2
+
+### Patch Changes
+
+- fix repeaterGroup item issue delete issue
+
 ## 0.2.1
 
 ### Patch Changes